% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_co2_vaqao.R
\name{sim_co2_vaqao}
\alias{sim_co2_vaqao}
\title{Genración de datos de CO2 en l'eliana}
\usage{
sim_co2_vaqao(id_dispositivo)
}
\arguments{
\item{id_dispositivo}{}
}
\value{
json
}
\description{
Genración de datos de CO2 en l'eliana
}
\examples{
 sim_co2_vaqao(id_dispositivo)

}
